Update tests and code with comments
