+++ tests/unit/commands/local/lib/test_sam_api_provider.py
@@ -7,28 +7,9 @@ from nose_parameterized import parameterized
 
 from six import assertCountEqual
 
-from samcli.commands.local.lib.cf_api_provider import CFApiProvider
 from samcli.commands.local.lib.api_provider import ApiProvider, SamApiProvider
 from samcli.commands.local.lib.provider import Api
 from samcli.commands.validate.lib.exceptions import InvalidSamDocumentException
-
-
-class TestSamApiProvider_init(TestCase):
-
-    @patch.object(ApiProvider, "_extract_apis")
-    @patch("samcli.commands.local.lib.api_provider.SamBaseProvider")
-    def test_provider_with_valid_template(self, SamBaseProviderMock, extract_api_mock):
-        extract_api_mock.return_value = {"set", "of", "values"}
-
-        template = {"Resources": {"a": "b"}}
-        SamBaseProviderMock.get_template.return_value = template
-
-        provider = ApiProvider(template)
-
-        self.assertEquals(len(provider.apis), 3)
-        self.assertEquals(provider.apis, set(["set", "of", "values"]))
-        self.assertEquals(provider.template_dict, {"Resources": {"a": "b"}})
-        self.assertEquals(provider.resources, {"a": "b"})
 
 
 class TestSamApiProviderWithImplicitApis(TestCase):
@@ -493,7 +474,7 @@ class TestSamApiProviderWithExplicitApis(TestCase):
             provider = ApiProvider(template)
             assertCountEqual(self, self.input_apis, provider.apis)
 
-    @patch("samcli.commands.local.lib.cf_base_api_provider.SamSwaggerReader")
+    @patch("samcli.commands.local.lib.cfn_base_api_provider.SamSwaggerReader")
     def test_with_swagger_as_both_body_and_uri(self, SamSwaggerReaderMock):
         body = {"some": "body"}
         filename = "somefile.txt"
@@ -957,211 +938,6 @@ class TestSamApiProviderWithExplicitAndImplicitApis(TestCase):
         assertCountEqual(self, expected_apis, provider.apis)
 
 
-class TestSamApiProviderwithApiGatewayRestApi(TestCase):
-
-    def setUp(self):
-        self.binary_types = ["image/png", "image/jpg"]
-        self.input_apis = [
-            Api(path="/path1", method="GET", function_name="SamFunc1", cors=None),
-            Api(path="/path1", method="POST", function_name="SamFunc1", cors=None),
-
-            Api(path="/path2", method="PUT", function_name="SamFunc1", cors=None),
-            Api(path="/path2", method="GET", function_name="SamFunc1", cors=None),
-
-            Api(path="/path3", method="DELETE", function_name="SamFunc1", cors=None)
-        ]
-
-    def test_with_no_apis(self):
-        template = {
-            "Resources": {
-
-                "Api1": {
-                    "Type": "AWS::ApiGateway::RestApi",
-                    "Properties": {
-                    },
-
-                }
-            }
-        }
-
-        provider = ApiProvider(template)
-
-        self.assertEquals(provider.apis, [])
-
-    def test_with_inline_swagger_apis(self):
-        template = {
-            "Resources": {
-
-                "Api1": {
-                    "Type": "AWS::ApiGateway::RestApi",
-                    "Properties": {
-                        "Body": make_swagger(self.input_apis)
-                    }
-                }
-            }
-        }
-
-        provider = ApiProvider(template)
-        assertCountEqual(self, self.input_apis, provider.apis)
-
-    def test_with_swagger_as_local_file(self):
-        with tempfile.NamedTemporaryFile(mode='w') as fp:
-            filename = fp.name
-
-            swagger = make_swagger(self.input_apis)
-            json.dump(swagger, fp)
-            fp.flush()
-
-            template = {
-                "Resources": {
-
-                    "Api1": {
-                        "Type": "AWS::ApiGateway::RestApi",
-                        "Properties": {
-                            "BodyS3Location": filename
-                        }
-                    }
-                }
-            }
-
-            provider = ApiProvider(template)
-            assertCountEqual(self, self.input_apis, provider.apis)
-
-    def test_body_with_swagger_as_local_file_expect_fail(self):
-        with tempfile.NamedTemporaryFile(mode='w') as fp:
-            filename = fp.name
-
-            swagger = make_swagger(self.input_apis)
-            json.dump(swagger, fp)
-            fp.flush()
-
-            template = {
-                "Resources": {
-
-                    "Api1": {
-                        "Type": "AWS::ApiGateway::RestApi",
-                        "Properties": {
-                            "Body": filename
-                        }
-                    }
-                }
-            }
-            self.assertRaises(Exception, ApiProvider, template)
-
-    @patch("samcli.commands.local.lib.cf_base_api_provider.SamSwaggerReader")
-    def test_with_swagger_as_both_body_and_uri(self, SamSwaggerReaderMock):
-        body = {"some": "body"}
-        filename = "somefile.txt"
-
-        template = {
-            "Resources": {
-
-                "Api1": {
-                    "Type": "AWS::ApiGateway::RestApi",
-                    "Properties": {
-                        "BodyS3Location": filename,
-                        "Body": body
-                    }
-                }
-            }
-        }
-
-        SamSwaggerReaderMock.return_value.read.return_value = make_swagger(self.input_apis)
-
-        cwd = "foo"
-        provider = ApiProvider(template, cwd=cwd)
-        assertCountEqual(self, self.input_apis, provider.apis)
-        SamSwaggerReaderMock.assert_called_with(definition_body=body, definition_uri=filename, working_dir=cwd)
-
-    def test_swagger_with_any_method(self):
-        apis = [
-            Api(path="/path", method="any", function_name="SamFunc1", cors=None)
-        ]
-
-        expected_apis = [
-            Api(path="/path", method="GET", function_name="SamFunc1", cors=None),
-            Api(path="/path", method="POST", function_name="SamFunc1", cors=None),
-            Api(path="/path", method="PUT", function_name="SamFunc1", cors=None),
-            Api(path="/path", method="DELETE", function_name="SamFunc1", cors=None),
-            Api(path="/path", method="HEAD", function_name="SamFunc1", cors=None),
-            Api(path="/path", method="OPTIONS", function_name="SamFunc1", cors=None),
-            Api(path="/path", method="PATCH", function_name="SamFunc1", cors=None)
-        ]
-
-        template = {
-            "Resources": {
-                "Api1": {
-                    "Type": "AWS::ApiGateway::RestApi",
-                    "Properties": {
-                        "Body": make_swagger(apis)
-                    }
-                }
-            }
-        }
-
-        provider = ApiProvider(template)
-        assertCountEqual(self, expected_apis, provider.apis)
-
-    def test_with_binary_media_types(self):
-        template = {
-            "Resources": {
-
-                "Api1": {
-                    "Type": "AWS::ApiGateway::RestApi",
-                    "Properties": {
-                        "Body": make_swagger(self.input_apis, binary_media_types=self.binary_types)
-                    }
-                }
-            }
-        }
-
-        expected_binary_types = sorted(self.binary_types)
-        expected_apis = [
-            Api(path="/path1", method="GET", function_name="SamFunc1", cors=None,
-                binary_media_types=expected_binary_types),
-            Api(path="/path1", method="POST", function_name="SamFunc1", cors=None,
-                binary_media_types=expected_binary_types),
-
-            Api(path="/path2", method="PUT", function_name="SamFunc1", cors=None,
-                binary_media_types=expected_binary_types),
-            Api(path="/path2", method="GET", function_name="SamFunc1", cors=None,
-                binary_media_types=expected_binary_types),
-
-            Api(path="/path3", method="DELETE", function_name="SamFunc1", cors=None,
-                binary_media_types=expected_binary_types)
-        ]
-
-        provider = ApiProvider(template)
-        assertCountEqual(self, expected_apis, provider.apis)
-
-    def test_with_binary_media_types_in_swagger_and_on_resource(self):
-        input_apis = [
-            Api(path="/path", method="OPTIONS", function_name="SamFunc1"),
-        ]
-        extra_binary_types = ["text/html"]
-
-        template = {
-            "Resources": {
-
-                "Api1": {
-                    "Type": "AWS::ApiGateway::RestApi",
-                    "Properties": {
-                        "BinaryMediaTypes": extra_binary_types,
-                        "Body": make_swagger(input_apis, binary_media_types=self.binary_types)
-                    }
-                }
-            }
-        }
-
-        expected_binary_types = sorted(self.binary_types + extra_binary_types)
-        expected_apis = [
-            Api(path="/path", method="OPTIONS", function_name="SamFunc1", binary_media_types=expected_binary_types),
-        ]
-
-        provider = ApiProvider(template)
-        assertCountEqual(self, expected_apis, provider.apis)
-
-
 class TestSamStageValues(TestCase):
 
     def test_provider_parse_stage_name(self):
@@ -1347,157 +1123,6 @@ class TestSamStageValues(TestCase):
         self.assertIn(api1, result)
         self.assertIn(api2, result)
         self.assertIn(api3, result)
-
-
-class TestApiProviderSelection(TestCase):
-    def test_api_provider_sam_api(self):
-        resources = {
-            "TestApi": {
-                "Type": "AWS::Serverless::Api",
-                "Properties": {
-                    "StageName": "dev",
-                    "DefinitionBody": {
-                        "paths": {
-                            "/path": {
-                                "get": {
-                                    "x-amazon-apigateway-integration": {
-                                        "httpMethod": "POST",
-                                        "type": "aws_proxy",
-                                        "uri": {
-                                            "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31"
-                                                       "/functions/${NoApiEventFunction.Arn}/invocations",
-                                        },
-                                        "responses": {},
-                                    },
-                                }
-                            }
-
-                        }
-                    }
-                }
-            }
-        }
-
-        provider = ApiProvider.find_correct_api_provider(resources)
-        self.assertTrue(isinstance(provider, SamApiProvider))
-
-    def test_api_provider_sam_function(self):
-        resources = {
-            "TestApi": {
-                "Type": "AWS::Serverless::Function",
-                "Properties": {
-                    "StageName": "dev",
-                    "DefinitionBody": {
-                        "paths": {
-                            "/path": {
-                                "get": {
-                                    "x-amazon-apigateway-integration": {
-                                        "httpMethod": "POST",
-                                        "type": "aws_proxy",
-                                        "uri": {
-                                            "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31"
-                                                       "/functions/${NoApiEventFunction.Arn}/invocations",
-                                        },
-                                        "responses": {},
-                                    },
-                                }
-                            }
-
-                        }
-                    }
-                }
-            }
-        }
-
-        provider = ApiProvider.find_correct_api_provider(resources)
-
-        self.assertTrue(isinstance(provider, SamApiProvider))
-
-    def test_api_provider_cloud_formation(self):
-        resources = {
-            "TestApi": {
-                "Type": "AWS::ApiGateway::RestApi",
-                "Properties": {
-                    "StageName": "dev",
-                    "Body": {
-                        "paths": {
-                            "/path": {
-                                "get": {
-                                    "x-amazon-apigateway-integration": {
-                                        "httpMethod": "POST",
-                                        "type": "aws_proxy",
-                                        "uri": {
-                                            "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31"
-                                                       "/functions/${NoApiEventFunction.Arn}/invocations",
-                                        },
-                                        "responses": {},
-                                    },
-                                }
-                            }
-
-                        }
-                    }
-                }
-            }
-        }
-
-        provider = ApiProvider.find_correct_api_provider(resources)
-        self.assertTrue(isinstance(provider, CFApiProvider))
-
-    def test_multiple_api_provider_cloud_formation(self):
-        resources = {
-            "TestApi": {
-                "Type": "AWS::ApiGateway::RestApi",
-                "Properties": {
-                    "StageName": "dev",
-                    "Body": {
-                        "paths": {
-                            "/path": {
-                                "get": {
-                                    "x-amazon-apigateway-integration": {
-                                        "httpMethod": "POST",
-                                        "type": "aws_proxy",
-                                        "uri": {
-                                            "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31"
-                                                       "/functions/${NoApiEventFunction.Arn}/invocations",
-                                        },
-                                        "responses": {},
-                                    },
-                                }
-                            }
-
-                        }
-                    }
-                }
-            },
-            "OtherApi": {
-                "Type": "AWS::Serverless::Api",
-                "Properties": {
-                    "StageName": "dev",
-                    "DefinitionBody": {
-                        "paths": {
-                            "/path": {
-                                "get": {
-                                    "x-amazon-apigateway-integration": {
-                                        "httpMethod": "POST",
-                                        "type": "aws_proxy",
-                                        "uri": {
-                                            "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31"
-                                                       "/functions/${NoApiEventFunction.Arn}/invocations",
-                                        },
-                                        "responses": {},
-                                    },
-                                }
-                            }
-
-                        }
-                    }
-                }
-            }
-        }
-
-        provider = ApiProvider.find_correct_api_provider(resources)
-        self.assertTrue(isinstance(provider, CFApiProvider))
 
 
 def make_swagger(apis, binary_media_types=None):
